wip: needs refactor
