+++ src/main/java/io/confluent/idesidecar/restapi/kafkarest/controllers/AdminClientService.java
@@ -1,17 +1,9 @@
 package io.confluent.idesidecar.restapi.kafkarest.controllers;
 
-import static io.confluent.idesidecar.restapi.util.RequestHeadersConstants.CONNECTION_ID_HEADER;
-
 import io.confluent.idesidecar.restapi.connections.ConnectionStateManager;
-import io.vertx.core.http.HttpServerRequest;
 import jakarta.enterprise.context.ApplicationScoped;
-import jakarta.enterprise.context.RequestScoped;
-import jakarta.enterprise.context.SessionScoped;
-import jakarta.enterprise.inject.Disposes;
 import jakarta.inject.Inject;
-import jakarta.ws.rs.Produces;
 import java.util.Properties;
-import org.apache.kafka.clients.admin.AdminClient;
 
 @ApplicationScoped
 public class AdminClientService {
